requirements.txt
fastapi
uvicorn[standard]
python-multipart
langchain
langchain-core
langchain-openai
langchain-google-genai
python-dotenv
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
gspread
gspread
fastapi
uvicorn[standard]
pydantic[email]